backend/requirements.txt
fastapi
uvicorn
qdrant-client
python-dotenv
numpy
openai